[FEAT] FeedLocation 수정 : 위도와 경도 데이터 타입 double로 변경

+++ src/main/java/com/example/trip/dto/response/FeedLocationResponseDto.java
@@ -15,8 +15,8 @@ public class FeedLocationResponseDto {
         private Long placeId;
         private String placeName;
         private String placeAddress;
-        private Long latitude;
-        private Long longitude;
+        private double latitude;
+        private double longitude;
         private List<FeedDetailLocResponseDto.GetFeedDetailLoc> feedDetailLoc;
